+++ RecoLocalCalo/HcalRecAlgos/interface/HcalCorrectionFunctions.h
